crosshatch: uevent: change fingerprint sysfs node owner to system
Without this, ls -l shows the sysfs nodes exposed by the FPC kernel
driver and used by the fingerprint HAL located at:

/sys/devices/platform/soc/soc:fp_fpc1020/*

as owned by root.  This makes is hard for init to start the HAL as a
service as the system user, since it can't read nodes owned by root.

Bug: 64818579
Change-Id: Ibcaf97f7b2e8af8273067cde9e35c94283a2ff9d