Add incall music to crosshatch mixer paths and policy
This change adds the paths for crosshatch and blueline for the incall music
use case. The path is which is used is incall-music-uplink.
A new mix port is added for the incall-music usecase which matches the
AUDIO_OUTPUT_FLAG_INCALL_MUSIC which will be introduced as part of a
different change.
This change was tested together with a set of additional changes to the
audio framework nd HAL which introduces the ability to play pcm audio to the uplink stream.

Bug: 69973354.
Test: Tested Manually.
Change-Id: I5e864c010a8f1e3cf6f6b51da10195027a6cd79e