WiFi: DLKM load once unload never for crosshatch
Current WiFi driver implementation for crosshatch devices is based on a
pure DLKM mode (Dynamically Loaded Kernel Module). This means, the
driver is loaded/unloaded by the WiFi HAL
This requires the WiFi hal to have admin access on loading/unloading
kernel modules which may result in some security risks

This Commit switches to the mode where the wifi driver is loaded via the
init scripts at boot and not unloaded throughout the operation. Hence no
extra access rights are required for the wifi hal

Bug: 69554076
Test: Manual Test

Change-Id: I5392eaae648b7e4d9c000b291dc51927c6f490e0
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>