Low Power Monitor system property security policy for crosshatch
Adds a security policy allowing ConnectivityMonitor to modify the new
system property controlling startup of the Low Power Monitor service.

This has not been tested on device due to availability, but the issue
and the policy to address it are the same as found on wahoo, and
details of the specific failure can be found on that CL (ag/3502344).
This feature is currently disabled for crosshatch and will remain that
way until system property modification has been tested.

Bug:35955665
Test: Still in progress as part of the main code changes in a different
CL

Change-Id: If9fc263275f5ab9dec2b205f93736dd79704ea59