Wifi: Disable forcing 11n for SAP
Current Implementation does not support 11ac for SAP mode.
This was due to some certification limitation. Since this limitation is
no longer valid, we need to re-enable the use of 11AC for SAP

This commit resets the configuration in WCNSS_qcom_cfg.ini
by setting gSapForce11NFor11A to zero

Bug: 72664404
Test: Manual

Change-Id: Ia2925de54e5f5a1bd79a0a25629d96b7e834ea13
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>