haptics: cs40l20: add firmware with f0 tracking support
This patch updates the firmware that supports f0 tracking as well as LRA
resistance measurement. This new firmware requires a new waveform
library format that's incompatible with the old version. We'll remove
the waveform library from the device for now to prevent it from for the
driver to fail since currently the waveform is not being used.

Bug: 70630396
Test: Follow instructions on go/b1c1-haptics-cal
Change-Id: I6435b47f27743c4c05ceb5f31201cab431cb0df2
Signed-off-by: David Lin <dtwlin@google.com>