Fix SELinux denials that occur a lot in enforcing mode.
Many of these denials occur a lot in enforcing mode, slowing down the device.

The rcs service was not being labeled, which was causing a crash loop, so I moved the definition to the private version of the file.

The netmgrd changes were copied from Wahoo.  It was also accessing /sys/devices/virtual/net/rmnet_ipa0/queues/rx-0/rps_cpus, so I copied Wahoo's labeling of that directory.

Sensors was trying to access the two devices many times.  The file it was accessing was /sys/devices/soc0/hw_platform.

I'm not sure about the turing directory change, but all of the 15 or so processes I saw accessing it shared only sysfs_msm_subsys, and there are some kernel commits that mention Turing and msm (e.g., kernel commit 2c67f5cc), so it seems reasonable.

Bug: 72643420
Test: Boot device and test wifi.  Verify that the denials are gone.
Change-Id: I1c32ecb18d9ab8b6852d78d00b4d0c683c3a5816