Fix edge sense in enforcing mode.
This allows edge sense to work in enforcing mode.  Commit e7d3e475
added a temporary script that initializes edge sense and is called by
init.qcom.devstart.sh.  This adds the permissions required for that
script to work.  We can remove them when we remove the script.

Bug: 72643420
Test: Booted device in enforcing mode and launched edge sense by
squeezing the sides.

Change-Id: I301c3b8a873ce694a9409f297be51fadfe984859