Allowing GmsCore to register DnD rules
This is necessary because the driving DnD rules will be migrated from
Pixel Ambient Services to GmsCore. Pre-granting this permission is
necessary to avoid a feature regression when compared to Pixel 2.
Access for intelligence.sense will be revoked once the system image
contains a build of GmsCore that includes the driving DnD rule.

Bug: 74115747
Test: N/A
Change-Id: Iecccc8530691af6aa74d989a1b11908a2af22583