Add SettingsIntelligence to Crosshatch DND access.
This is set in the nexus_overlay and no differences existed until
GMSCore was added to the nexus_overlay. GMSCore only exists in the Wahoo
overlay and the overlay currently crushes our config which applies to
all Google devices. By re-adding it to this overlaid overlay, we gain
our powers back.

Bug: 72663665
Test: Manual
Change-Id: Ia498d4a3c35ab24af3a99b29f6c31562a76e0039