haptics: fix issue with ReDC calculation
This change fixes an issue with the script that generates the waveform
table was using an incorrect equation. The scaling factor listed in the
firmware release notes for ReDC (5.687) has been incorrect since the
beginning, and should be 5.857. Therefore the formulae for redc_measured
and redc_stored change as follows:

ReDC (ohms) = redc_measured / 2^17 * 5.857
redc_stored = ReDC (ohms) / 5.857 * 2^17

Bug: 70630396
Test: perform cal and manual verification
Change-Id: I08eaa44c5d7c758b682d1ef1f8423677f40350a8
Signed-off-by: David Lin <dtwlin@google.com>