DO NOT MERGE: Allow RILD to read xt_qtaguid iface stats on b1c1
The RILD process need to get the per iface stats from
proc/net/xt_qtaguid/iface_stat_fmt file. It is too risky to change them
to using the Java level API for P release so we have to grant them this
permission for another year. In Q they will switch to the public API so
this change only goes into pi-dev.

Bug: 68774956
Test: device boot without selinux violation
Change-Id: Ib86916951cb8f340bfef55814ae8c4fef0f51338