2018 PowerHal support for Easel stats
Now that real Easel power state stats are being exported by the driver,
this CL replaces the previous Easel "SyntheticSleep" stat with support
for the real stats.  Specifically:
* support is added for reading and parsing stats for the three Easel
  power states
* support for the temporary SyntheticSleep power state is removed
* sepolicy is added to allow power_hal to read the easel stats file:
  /d/mnh_sm/power_stats
* sepolicy is removed that allowed power_hal to read the easel
  state (not stats) file: /sys/devices/virtual/misc/mnh_sm/state

Note that the Easel stats debugfs node is currently created with 0400
permissions by the driver, which makes it inaccessible even with the
correct sepolicy in place.  The other CL in this topic modifies the
driver to create the debugfs node with 0444 permissions, which matches
the permissions on other system/subsystem power stats files.

Bug: 36576572
Test: Checked audit logging iteratively from "no new sepolicy" to the
changes in this CL to ensure that only minimal additions are introduced.
Test: Ran image on blueline device (with kernel built to include
associated driver change), allowed device to drain for 30 minutes, then
exampled 'dumpsys batterystats' output and verified that Easel power
states' stats wwere present.

Change-Id: I2d386f5e24a24e36a156aed5ac17d2ef8a07e449