audio: correct the acdb id of usb playback
The acdb id of usb-headphones playback is chosen to be that chosen of
voice usb-headphones. However, the playback does not require any voice
tuning hence it should be no difference of acdb selection between
usb-headphones and usb-headset with playback scenario. Then, remove the
definition to set back to the qcom default.

Bug: 78444023
Test: manual audio test
Change-Id: If3571897aa1db163fb0c13b8d1a8be3f9d43b9f9
Signed-off-by: HW Lee <hwlee@google.com>