WiFi: Add SAR body sensor support to crosshatch
This commit adds the configurations for the support of body/hand/head
sensor to WiFi SAR in crosshatch devices.

Note: The configurations for the support of SAR and SAR sensor are
currently set to false. This is awaiting some dependencies to be resolved.
A bug is filed to bring both configuration to true once these
dependencies are resolved (b/109931816).

Bug: 65174506
Test: Run Wifi unit test suite
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: I3ff90074c4057958b4ce3c2216198c32e0f6175b
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>