thermal: recommend throttling at low temps
Based on the State-of-Charge, the temperature, and the cycle count,
recommend throttling at low temperatures as well. Cold, aged, and
near-empty batteries cannot supply as much peak current. This change
will recommend mitigation (currently, turning off the camera flash)
when the conditions of the LookupTable are met.
These conditions are only activated in subzero (deg C) conditions.

Test: VtsHalThermalV1_0TargetTest pass 3/3
Test: VtsHalThermalV1_1TargetTest pass 1/1
Test: new PtsThermalHalTestCases sanity check
Test: PtsThermalHalTestCases pass 6/6
Test: manually test via using faked battery sysfs files and kernel
      with CONFIG_THERMAL_EMULATION, see threshold reporting through
      media.camera dumpsys.
Test: lshal debug "android.hardware.thermal@1.1::IThermal/default"
Bug: 80202800

Change-Id: I318e77e5d040ba438c8ead93b49d6f00a65a9de1