Wifi: Remove early setting of wlan.driver.status
In current implementation, the wlan.driver.status property is
set to OK inside the init script once driver is loaded.
This is too early, and driver may not be ready yet by that time.
This resulted in breaking the feature to set driver and FW version
properties.

This commit removes this early setting, and relies on the HAL to
set it when driver is ready.

Bug: 76220544
Test: Boot the device and make sure the version properties are set
Test: adb shell getprop sys.wlan.driver.version
Test: adb shell getprop sys.wlan.firmware.version

Change-Id: I4a1921f13c9b8077c78320fca593f53352cb922c
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>