haptics: update the PWLE timing for the delay
This patch updates the PWLE timing in delay to better sync the ringtone
playback with the latest build that fixes the issue with inconsistent
delay when repeating the ringtone. This change also adjust the amplitude
of index #2 and #3 for the click effect to match with the one from the
UX design.

Bug: 75947705
Test: manual ringtone haptics test
Change-Id: Ic3042047462bb189fc91390c4ec4601c54ca2b63
Signed-off-by: David Lin <dtwlin@google.com>