Change VR brightness overlay of Blueline and Crosshatch
VR brightness on blueline and crosshatch are changed so that the latest
dogfood devices shall be able to open vr with correct brightness.

This patch provides a fix based on experiments.

Test: build blueline/crosshatch and run a vr app. Correct brightness
shall be obtained by sliding the brightness controller UI, so that the
brightness bug in HAL can be bypassed.
Bug: 80476493
Bug: 80140391

Change-Id: Ia49939e78e6231ff2766a697bd535ee20ddd6c75