Separate thermalHAL scale multiplier for each sensor.
ThermalHAL used to use a global variable to keep track of the scale
multiplier for all sensors. It was the case that all sensors had the
same scale multipler earlier, but it may not be the case in the future.
This CL separates the scale multiplier for each sensor and adds it to
the kValidThermalSensorTypeMap. The multipliers used in parsing the
configs are still kept the same.

Bug: 76443195
Test: pts-tradefed run pts -m PtsThermalHalTestCases
Change-Id: Ib2274ee5fc947cafe64e198001d3c33e1a761609