Change thermalHAL to use threshold from skin monitor config.
ThermalHAL used to use the minimum threshold seen for skin sensor as its
threshold. This changes thermalHAL to use only the min threshold from
the [SKIN-MONITOR] section of the config. This change adds a rule_name
to SensorConfig in ThermalConfigParser.cpp and changes the logic when
setting the skin sensor threshold. Note that this will only affect
clients that use the throttling value from the getTemperatures() api of
ThermalHAL1.0.

Bug: 80306322
Test: pts-tradefed run pts -m PtsThermalHalTestCases
Test: manual - boot the device and check thermalHAL loglines from logcat
Test: manual - manually do getTemperatures() and check threshold for ss
Change-Id: Ic8fcf592932fa0795fffaab0fed47d28844e0b5f