crosshatch: pixelstats: hold a ref to DropDetect
DropDetect works by spawning a thread which holds
a reference to the DropDetect object.
If that thread exits, it clears it's sp to itself
during shutdown, which ends up trying to destruct the
thread that's calling the destructor.

Hold a reference to DropDetect in the pixelstats service
to make sure this doesn't happen.

Test: kill pixelstats with various signals, it always
went down cleanly.
I can't repro the original failure, though this is
cleaner.

Bug: 111439902
Change-Id: Iaf78e0e708f674f7caffd197fd06bbb315b55044