thermal: Use explicit disable for USB overheat mitigation
Right now, the overheat mitigation disable mechanism doesn't
persist if the device is in high temperatures during boot.
It also doesn't do anything if the device is already in
USB overheat mitigation. This change uses a module parameter
that is specifically reserved for testing and will disable
overheat mitigation even if it has already kicked off.

Bug: 111752140
Change-Id: I8d0dc63a39184a000801acc156d9899abae21cc7
Signed-off-by: Maggie White <maggiewhite@google.com>