Add init.firstboot.sh, for factory use case
In the factory they need an automated way to shut down the device after
setupwizard finishes, but we also don't want to let them talk to Android
over USB on a user / locked device.

When bootmode is "firstboot" (Set by bootloader), the device will wait
for USB cable to be disconnected, then shut down.

Bug: 110896488
Change-Id: I9b831ef37e0d8d3f84a315c4d23b8258702b95fc