PowerHAL: add debug support for low power stats
The HAL API includes a "debug" interface that can be invoked from the
command line to get debug output for a particular HAL.  While the
existing PowerHAL implementation includes support for the "debug"
interface, it only includes data for the performance hinting side of the
HAL.  To facilitate device troubleshooting (not just of the HAL itself,
but of power drain), this change expands that implementation to also
include the output from calls to getPlatformLowPowerStats() and
getSubsystemLowPowerStats().

On existing in-market devices where this support is present, the command
    lshal debug android.hardware.power@1.1::IPower/default
will produce the "debug" output.

Note that the standalone platform and subsystem stats display functions
referenced by this CL are being added to a new library in a different CL
(see topic) so they can be shared across device families where
appropriate to reduce code duplication.

Bug: 116750864
Test: Ran code on a device and used 'lshal debug ...' to generate output.
Test: Added temporary test code (triggered by appending " test" to the
command above) to cover all base cases of data being present absent
(test code is included in first uploaded patchset on this CL).

Change-Id: Id97534a2a9fe05bbca614c763c907daea37e9723