Stop kernel from loading firmware directly
On crosshatch, we store firmware files in the following two locations:

/vendor/firmware_mnt/image/ and
/vendor/firmware/

Depending on the location, we currently use two different mechanisms to
load firmware files. For files in /vendor/firmware/, we have the kernel
load files directly using the following kernel command line argument:

BOARD_KERNEL_CMDLINE += firmware_class.path=/vendor/firmware

Files in /vendor/firmware/ are loaded by the user mode helper ueventd
that the kernel communicates with via NETLINK.

Having two different ways to load firmware is not desireable.  Depending
on whether ueventd or the kernel accesses a firmware file, different
access controls need to be adjusted including a modified sepolicy. In
the case where the kernel loads firmware files directly, files are
sometimes accessed with UID 0 (i.e. root) and sometimes with other UIDs
depending on which context the kernel is in.

Bug: 117290503
Change-Id: Ic21cae665b0e360d9ff911f2f99e9b2c24c3d371