Remove AAPT config from _mainline targets
These options are meant to save space on the system image, but
simultaneously make the system image less generic than we would
like it to be.

Experiments show the savings on /system are <20 MB and /product 11 MB.
Remove these flags for the mainline targets, as they're not worth the
hassle.

Bug: 118676832
Test: make
Change-Id: Ifb7743cb157fbbfe144590d912392530961835cd