Prepare build flags for spanning the super partition across multiple block devices.
This patch enables the following when
PRODUCT_RETROFIT_DYNAMIC_PARTITIONS is true.

This adds BOARD_SUPER_PARTITION_BLOCK_DEVICES, which must contain a list
of the (non-A/B suffixed) partitions that will comprise the super
partition. It is only intended for devices which cannot have a partition
named "super". For each entry, there must be a
BOARD_SUPER_PARTITION_x_DEVICE_SIZE variable defined with the exact size
of that partition (not its image size). The sum of these sizes must be
equal to BOARD_SUPER_PARTITION_SIZE.

Bug: 116802789
Test: device with PRODUCT_RETROFIT_DYNAMIC_PARTITIONS builds
Change-Id: Ib9b751a81029753221ccd18db374476f8f57e337