Fixing music detector audio buffer size
This is the same fix that was done in master, see ag/5504798
now adding it to pi-dev so it can go out with a maintenance release.

This cl changes the buffer size of the bridge-buffer modules
for the music detector on the WDSP so that they can hold 8
seconds of mu-law encoded audio (ie. 0x1F400 or 128000 bytes).

Test: manually tested on crosshatch, verified that music detection and
hotword work as expected.
Bug: 118797037
Change-Id: Ibd3b3a2578e67ccb467a78601098ae91cb0406ec
Merged-In: I8d35c937d6adc5781a83b3b3c8ce99969219d0b5