Remove WAPPushManager from crosshatch devices
I believe this is unused. Nothing related to it appears to be
included in the non aosp_ variants of these devices.

Also minor tweaks to the layout of this file, to make changes
less likely to conflict with internal.

Test: boot aosp_blueline
Change-Id: I47ffb360bf5422d1e26a03367a2b2179ebee67b1
Merged-In: I47ffb360bf5422d1e26a03367a2b2179ebee67b1