Resume-on-Reboot: use uevent to chown
Previously the device was set via the HAL .rc file, but we now allow this device
to change names to accomodate other board types that use pmem. Setting this
in the HAL .rc file now causes SELinux errors on those platforms.

Switch to using the ueventd.rc method to avoid this SELinux problem.

Bug: 63928581
Bug: 146400078
Test: atest VtsHalRebootEscrowTargetTest
Change-Id: I6c3b40fdf1a3e7a60bdab4dacaea60fa6df06ec1